Object and controller classes for the application
The StudentIFATCard is the object oriented class for the IFAT card and it contains the grading algorithm. The IFATController class contains the main functional methods for the application such as the image detection, and load cards methods. The IFATCardGrading class is just to test the grading algorithm from the StudentIFATCard class.